Remove Application and ApplicationType entirely. Move front controller path property to Php options
