--- src/Form/Generator/ApplicationType.php
@@ -1,49 +1,0 @@
-<?php
-declare(strict_types=1);
-/**
- * Copyright 2016 Luis Alberto Pabón Flores
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-namespace App\Form\Generator;
-
-use Symfony\Component\Form\Extension\Core\Type\TextType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\Validator\Constraints\Length;
-use Symfony\Component\Validator\Constraints\NotBlank;
-use Symfony\Component\Validator\Constraints\NotNull;
-
-/**
- * Form for application options.
- */
-class ApplicationType extends AbstractGeneratorType
-{
-    /**
-     * Builds the form definition.
-     */
-    public function buildForm(FormBuilderInterface $builder, array $options): void
-    {
-        $builder
-            ->add('frontControllerPath', TextType::class, [
-                'label'       => 'Front controller path (relative to container workdir)',
-                'attr'        => ['placeholder' => 'Password for root user'],
-                'data'        => 'public/index.php',
-                'constraints' => [
-                    new NotBlank(),
-                    new NotNull(),
-                    new Length(min: 2, max: 128),
-                ],
-            ]);
-    }
-}
