+++ src/PHPDocker/Generator/Files/Dockerfile.php
@@ -48,7 +48,6 @@ class Dockerfile implements GeneratedFileInterface
         $data = [
             'phpVersion'       => $this->project->getPhpOptions()->getVersion(),
             'packages'         => array_unique($packages),
-            'applicationType'  => $this->project->getApplicationOptions()->getFrontControllerPath(),
             'dockerWorkingDir' => $this->project->getWorkingDirOptions()->getDockerWorkingDir(),
         ];
