--- src/PHPDocker/Project/ServiceOptions/Application.php
@@ -1,38 +1,0 @@
-<?php
-declare(strict_types=1);
-/*
- * Copyright 2021 Luis Alberto Pabón Flores
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-namespace App\PHPDocker\Project\ServiceOptions;
-
-/**
- * Describes a few things about the user's project.
- */
-class Application
-{
-    /**
-     * @param string $frontControllerPath Path to the app's front controller, relative to project root
-     */
-    public function __construct(private string $frontControllerPath)
-    {
-    }
-
-    public function getFrontControllerPath(): string
-    {
-        return $this->frontControllerPath;
-    }
-}
